+++ lib/cg/Window.cpp
@@ -2,12 +2,12 @@
 
 namespace CG {
 
-  Window::Window(float xmin, float ymin, float xmax, float ymax)
+  Window::Window(double xmin, double ymin, double xmax, double ymax)
     : _xmin(xmin), _ymin(ymin), _xmax(xmax), _ymax(ymax) {
 
   }
 
-  void Window::zoom(float step) {
+  void Window::zoom(double step) {
     if (_xmin + step >= _xmax - step || _ymin + step >= _ymax - step) {
       // Maximum zoom reached
       return;
